API-415 Support of OpenID Connect's RP Logout
Add Support of rp logout, optional feature from OpenID Connect
and managed in the Identity Server.
 After init configuration of the SSO JS, you have only to call signOut on the
AsmodeeNet module to execute all the dialog with the IS, and you can pass in callback
the method that you want to do you if all is ok. (so on, delete you user session)
- Update Code
- Update Readme
- Update example
- Update version number => 0.2.0

https://daysofwonder.atlassian.net/browse/API-415